Use new implementation of stack from Pandas 2.1.0
In preparation for Pandas 3.0, a new and improved implementation of the stack() method was introduced.
There is a feature flag, future_stack, which, when set to True, enables the new implementation to be used.
If the flag is not set, a warning will be emitted.

More information about the implementation can be found here:
https://pandas.pydata.org/docs/whatsnew/v2.1.0.html

Crucially, the behavior regarding NA values has been improved. The new implementation will not create unnecessary
NA values, but instead, it will preserve them from the input without the need to use dropna=False. This change avoids
the potential issue of generating unnecessary NA values from multiple levels of stacking.